gpu: nvgpu: fix MISRA violations common.gr.ctx
Rule 10.6 forbids assignment of an 32-bit expression to a 64-bit target.
This patch fixes MISRA 10.6 errors in common/gr/ctx.c

Jira NVGPU-3808

Reviewed-on: https://git-master.nvidia.com/r/2152643
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/gpu/nvgpu/common/gr/ctx.c
@@ -467,8 +467,10 @@ void nvgpu_gr_ctx_patch_write(struct gk20a *g,
 			return;
 		}
 
-		nvgpu_mem_wr32(g, &gr_ctx->patch_ctx.mem, patch_slot, addr);
-		nvgpu_mem_wr32(g, &gr_ctx->patch_ctx.mem, patch_slot + 1U, data);
+		nvgpu_mem_wr32(g, &gr_ctx->patch_ctx.mem,
+						(u64)patch_slot, addr);
+		nvgpu_mem_wr32(g, &gr_ctx->patch_ctx.mem,
+						(u64)patch_slot + 1ULL, data);
 		gr_ctx->patch_ctx.data_count = nvgpu_safe_add_u32(
 						gr_ctx->patch_ctx.data_count, 1U);
 		nvgpu_log(g, gpu_dbg_info,
